agents/research/requirements.txt
httpx
google-genai
asyncpg
python-dotenv
beautifulsoup4
aiohttp